Clean up docs.

+++ codegen/nrbf/tests/array_single_primitive.rs
@@ -1,7 +1,7 @@
 use std::collections::BTreeMap;
 
 use const_str::concat_bytes;
-use nrbf::{data_type::Int32, RemotingMessage, Value};
+use nrbf::{RemotingMessage, Value};
 
 #[rustfmt::skip]
 const INPUT: &[u8] = concat_bytes!(
@@ -22,8 +22,8 @@ const INPUT: &[u8] = concat_bytes!(
 #[test]
 fn array_single_primitive() {
   let output = RemotingMessage::Value(
-    BTreeMap::from_iter([(Int32(1), Value::Array(vec![Value::Int64(67), Value::Int64(42)]))]),
-    Value::Ref(Int32(1)),
+    BTreeMap::from_iter([(1, Value::Array(vec![Value::Int64(67), Value::Int64(42)]))]),
+    Value::Ref(1),
   );
 
   assert_eq!(RemotingMessage::parse(INPUT), Ok(([].as_slice(), output)));
